tf-gcp-lakehouse/02_cloudfunctions/gcs_trigger_function/requirements.txt
# requirements.txt

requests
functions-framework
cloudevents
google-auth
google-api-python-client
